MK-1205 Fix logged error when serialize AuthorizationModel to ehcache
